apps/api/requirements.txt
# Web framework
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0

# Settings & validation
pydantic>=2.8.0
pydantic-settings>=2.5.0

# Auth (JWT + password hashing)
PyJWT>=2.9.0
bcrypt>=4.2.0

# Database (async runtime + sync for Alembic)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Document parsers (XML, CSV, EDIFACT — JSON uses stdlib)
lxml>=5.3.0
pandas>=2.2.0
pydifact>=0.1.7

# S3-compatible object storage (MinIO in dev, swappable for AWS S3 in prod)
minio>=7.2.0

# AI agents (LangChain + Anthropic Claude)
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# Observability (Phoenix + OpenTelemetry + OpenInference LangChain instrumentation)
arize-phoenix-otel>=0.6.0
openinference-instrumentation-langchain>=0.1.29
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Linting & formatting
ruff>=0.6.0

# Dev tools (sample generation — not used at runtime)
fpdf2>=2.8.0
